size_t for i

+++ function_pointers/1-array_iterator.c
@@ -11,7 +11,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i = 0;
+	size_t i = 0;
 
 	if (size == 0 || array == NULL || action == NULL)
 		return;
